Ajustando integração com pagamento

+++ checkout/src/main/java/br/com/fiap/postech/adjt/checkout/application/dto/PaymentCheckoutDTO.java
@@ -6,7 +6,7 @@ import org.springframework.validation.annotation.Validated;
 @Validated
 public record PaymentCheckoutDTO(
         String orderId,
-        Double amount,
+        Integer amount,
         String currency,
         PaymentDTO payment_method
 ) {
